requirements.txt
abacusai==0.9.2
boto3==1.14.20
botocore==1.17.20
GitPython==3.1.3
google-api-python-client==1.11.0
pyYAML==5.3.1
recombee-api-client==3.0.0
tqdm==4.47.0
xminds
